Working on common tools again for workflow webpage.
