--- ToolBox/README.md
@@ -1,3 +1,0 @@
-# OpsSpace Toolbox
-
-@todo Documentation of toolbox contents goes here
