+++ config/packagesList.txt
@@ -2,3 +2,4 @@ AAAOps
 SiteReadiness
 TransferTeam
 WmAgentScripts
+WorkflowWebTools
