+++ source/setup.rst
@@ -2,12 +2,12 @@ Setup script
 ============
 
 Use as a script
--------------
+---------------
 
 .. program-output:: ../setup.py
 
 Members
 -------
 
-.. automodule:: setup
+.. automodule:: OpsSpace.setup
    :members:
