--- source/utils.rst
@@ -1,5 +1,0 @@
-Utils module
-============
-
-.. automodule:: ToolBox.utils
-   :members:
